pyproject.toml
[project]
name = "mcp-mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.3",
    "mysql-connector-python>=9.4.0",
    "pydantic-ai>=0.7.3",
    "python-dotenv>=1.1.1",
]
